Add router-based workspaces and theme persistence
