+++ src/App.css
@@ -1,103 +1,115 @@
 .app {
-	min-height: 100vh;
-	background: var(--bg);
-	color: var(--text);
+        min-height: 100vh;
+        background: var(--bg);
+        color: var(--text);
 }
 
 .container {
-	max-width: 1200px;
-	margin: 0 auto;
-	padding: 24px;
-	display: flex;
-	flex-direction: column;
-	gap: 24px;
+        max-width: 1200px;
+        margin: 0 auto;
+        padding: 24px;
+        display: flex;
+        flex-direction: column;
+        gap: 24px;
 }
 
 .banner {
-	padding: 12px;
-	border-radius: 10px;
-	border: 1px solid var(--accent);
-	background: var(--accent-subtle);
-	color: var(--text);
+        padding: 12px;
+        border-radius: 10px;
+        border: 1px solid var(--accent);
+        background: var(--accent-subtle);
+        color: var(--text);
 }
 
 .grid-wrap {
-	display: grid;
-	grid-template-columns: repeat(3, minmax(0, 1fr));
-	gap: 24px;
+        display: grid;
+        grid-template-columns: repeat(3, minmax(0, 1fr));
+        gap: 24px;
 }
 
 @media (max-width: 1100px) {
-	.grid-wrap {
-		grid-template-columns: repeat(2, minmax(0, 1fr));
-	}
+        .grid-wrap {
+                grid-template-columns: repeat(2, minmax(0, 1fr));
+        }
 }
 
 @media (max-width: 760px) {
-	.grid-wrap {
-		grid-template-columns: 1fr;
-	}
+        .grid-wrap {
+                grid-template-columns: 1fr;
+        }
 }
 
 .palette-info {
-	opacity: 0.7;
+        opacity: 0.7;
 }
 
 .palette-info span {
-	color: var(--accent);
+        color: var(--accent);
 }
 
 .workspace-bar {
-	display: grid;
-	grid-template-columns: 1fr auto auto;
-	gap: 8px;
-	align-items: center;
+        display: flex;
+        flex-wrap: wrap;
+        gap: 8px;
+        align-items: center;
 }
 
 .workspace-create {
-	display: grid;
-	grid-template-columns: 1fr auto;
-	gap: 8px;
-	margin-top: 10px;
+        display: grid;
+        grid-template-columns: 1fr auto;
+        gap: 8px;
+        margin-top: 10px;
 }
 
-.workspace-bar-select,
+.workspace-link {
+        color: var(--text);
+        text-decoration: none;
+        border: 1px solid var(--accent);
+        border-radius: 8px;
+        padding: 6px 8px;
+}
+
+.workspace-link.active {
+        background: var(--accent);
+        color: #fff;
+}
+
 .workspace-bar-input {
-	background: transparent;
-	color: var(--text);
-	border: 1px solid var(--accent);
-	border-radius: 8px;
-	padding: 6px 8px;
+        background: transparent;
+        color: var(--text);
+        border: 1px solid var(--accent);
+        border-radius: 8px;
+        padding: 6px 8px;
 }
 
 .muted {
-	opacity: 0.6;
+        opacity: 0.6;
 }
 
 .accent {
-	color: var(--accent);
+        color: var(--accent);
 }
 
 .palette-royalViolet {
-	--bg: #1b1b2f;
-	--card: #2c2c3e;
-	--text: #f1f1f8;
-	--accent: #9d4edd;
-	--accent-subtle: #c77dff;
+        --bg: #1b1b2f;
+        --card: #2c2c3e;
+        --text: #f1f1f8;
+        --accent: #9d4edd;
+        --accent-subtle: #c77dff;
 }
 
 .palette-coralSpark {
-	--bg: #202124;
-	--card: #2f3136;
-	--text: #eeeeee;
-	--accent: #ff6b6b;
-	--accent-subtle: #ff9f80;
+        --bg: #202124;
+        --card: #2f3136;
+        --text: #eeeeee;
+        --accent: #ff6b6b;
+        --accent-subtle: #ff9f80;
 }
 
 .palette-neonCyan {
-	--bg: #0d1b2a;
-	--card: #1b263b;
-	--text: #e0e1dd;
-	--accent: #00b4d8;
-	--accent-subtle: #90e0ef;
+        --bg: #0d1b2a;
+        --card: #1b263b;
+        --text: #e0e1dd;
+        --accent: #00b4d8;
+        --accent-subtle: #90e0ef;
 }
