+++ src/firebase.js
@@ -1,0 +1,10 @@
+import { initializeApp } from "firebase/app";
+import { getAuth, GoogleAuthProvider } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+import { firebaseConfig } from "./firebase-config";
+
+const fbApp = initializeApp(firebaseConfig);
+export const db = getFirestore(fbApp);
+export const auth = getAuth(fbApp);
+export const provider = new GoogleAuthProvider();
+
