+++ src/main.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
 import "./index.css";
 import { registerSW } from "./registerSW.js";
@@ -7,7 +8,9 @@ import { registerSW } from "./registerSW.js";
 registerSW();
 
 ReactDOM.createRoot(document.getElementById("root")).render(
-	<React.StrictMode>
-		<App />
-	</React.StrictMode>,
+        <React.StrictMode>
+                <BrowserRouter>
+                        <App />
+                </BrowserRouter>
+        </React.StrictMode>,
 );
